Feat(alignment): add multiline label and dynamic shape size features
I made a significant improvements to several parts, including dynamic
shape width and height size, dynamic proximity in the X and Y axes,
label issue related to shape overflow, label starting points, and
multiline label issues. Additionally, I centered everything on a single
Cairo surface.

However, the X and Y offsets for each shape, as well as the paper's
width and height, are not algorithmically determined; they are based on
drizzle lines, which I recognize is not the most appropriate method.
Nevertheless, this approach has resulted in much cleaner shapes in this
commit, which better control over label text fitting.